requirements.txt
elevenlabs
python-dotenv
requests
flask
